Add LogFormatV2 for human readable output
This commit adds the LogFormatV2 struct and WriterSinkV2 sink for
transparently converting the old LogFormat to the human readable
LogFormatV2 when writing to the sink output (io.Writer).

This approach allows for the log format to be made human readable
using the existing public, which is not changed - all that is required
to make the logs human readable is to use the new sink.

[#155369100]

Signed-off-by: Charlie Vieth <cviethjr@pivotal.io>