Honor all relevant flags in lagerflags.New
Previously, lagerflags.New would honor only the logLevel flag. This change uses the ConfigFromFlags helper to extract all the relevant flags into one config and then passes it to NewFromConfig so that the invocations are automatically consistent.

Since the default value for the timeFormat package variable is 0, corresponding to FormatUnixEpoch, and since the default value for redactSecrets is false, this default configuration is consistent with the previous hard-coded behavior.