fix: slog.Handler log error messages rather than `{}`
When used as an slog.Handler, error objects may be specified as slog.Attr attributes. In general attributes are
converted into lager.Data{} maps, but there's a feature of logging lager.Data{} maps which is that error objects
(and potentially other objects) are rendered as `{}` rather than a more useful representation. This adds a special
case for errors so that they are now converted to strings before being placed in a lager.Data{} map.